#327 feat: Create subroutes for Client gotchis
