+++ src/pages/Client/ClientRoutes.tsx
@@ -14,10 +14,8 @@ import { EthersApi } from 'api';
 import { CommonUtils } from 'utils';
 
 import { ClientAccount } from './routes/ClientAccount';
-import { ClientBorrowed } from './routes/ClientBorrowed';
 import { ClientGotchis } from './routes/ClientGotchis';
 import { ClientInstallations } from './routes/ClientInstallations';
-import { ClientLendings } from './routes/ClientLendings';
 import { ClientRealm } from './routes/ClientRealm';
 import { ClientTickets } from './routes/ClientTickets';
 import { ClientWarehouse } from './routes/ClientWarehouse';
@@ -112,9 +110,7 @@ export function ClientRoutes() {
 
             <Routes>
                 <Route path='' element={<ClientAccount />} />
-                <Route path='gotchis' element={<ClientGotchis />} />
-                <Route path='lendings' element={<ClientLendings />} />
-                <Route path='borrowed' element={<ClientBorrowed />} />
+                <Route path='gotchis/*' element={<ClientGotchis />} />
                 <Route path='installations' element={<ClientInstallations />} />
                 <Route path='warehouse' element={<ClientWarehouse />} />
                 <Route path='tickets' element={<ClientTickets />} />
